refactor(test): extract the suite corpus helpers out of XmlRoundTripTest
Three test classes now need to walk the same corpus, and the details are
easy to get subtly wrong: opening a suite file without a system id is what
lets xml/issue2501/2501.xml resolve its external entity, and resolving the
doctype from the classpath is what keeps the tests off the network.

The shared entity resolver also narrows the one XmlRoundTripTest had: it
substituted the bundled DTD for every entity, which is harmless for
serialized output but would feed the DTD to a &params; reference.

No behaviour change.
